64-bit compatibility work

+++ std/bigint.d
@@ -318,10 +318,10 @@ public:
     }
     /// Number of significant uints which are used in storing this number.
     /// The absolute value of this BigInt is always < 2^(32*uintLength)
-    int uintLength() { return data.uintLength(); }
+    size_t uintLength() { return data.uintLength(); }
     /// Number of significant ulongs which are used in storing this number.
     /// The absolute value of this BigInt is always < 2^(64*ulongLength)
-    int ulongLength() { return data.ulongLength(); }
+    size_t ulongLength() { return data.ulongLength(); }
 
     /** Convert the BigInt to string, passing it to 'sink'.
      *
